Fix dm

+++ back/src/wschat/wschat.service.ts
@@ -43,34 +43,34 @@ export class WschatService {
     const user = this.onlineUserService.getUser(socketId);
     const room = await this.chatService.getRoomById(subRoom.roomId);
     const penalty = await this.penaltiesService.getRoomPenaltiesForUser(user.id, room.id);   
- 
-    if (penalty && penalty.type === PenaltyType.BAN) {
-      this.eventEmitter.emit('room.user.join', {
-        room: room,
-        user: user,
-        success: false,
-        message: 'Vous avez été banni de ce salon' + (penalty.timetype === 'PERM' ? '.' : ' jusqu\'au ' + penalty.endTime.toLocaleString('fr-FR'))
-      });
-      return ;
-    }
-
-    if (user) {
-      if (!await this.chatService.canJoin(room.id, subRoom.password)) {
+    
+    if (user && room) {
+      if (penalty && penalty.type === PenaltyType.BAN) {
         this.eventEmitter.emit('room.user.join', {
           room: room,
           user: user,
           success: false,
-          message: 'Mot de passe incorrect',
-         });
-         return ;
+          message: 'Vous avez été banni de ce salon' + (penalty.timetype === 'PERM' ? '.' : ' jusqu\'au ' + penalty.endTime.toLocaleString('fr-FR'))
+        });
+        return ;
       }
 
-      await this.chatService.addUsersToRoom(room.id, user.id);
-      this.eventEmitter.emit('room.user.join', {
-        room: room,
-        user: user,
-        success: true,
-      });
+        if (!await this.chatService.canJoin(room.id, subRoom.password)) {
+          this.eventEmitter.emit('room.user.join', {
+            room: room,
+            user: user,
+            success: false,
+            message: 'Mot de passe incorrect',
+          });
+          return ;
+        }
+
+        await this.chatService.addUsersToRoom(room.id, user.id);
+        this.eventEmitter.emit('room.user.join', {
+          room: room,
+          user: user,
+          success: true,
+        });
     }
   }
 
@@ -119,7 +119,7 @@ export class WschatService {
     const user = this.onlineUserService.getUser(socketId);
     const target = await this.userService.findOne(room.targetId);
     const roomToEject = await this.chatService.getRoomById(room.roomId);
-    if (user && target.id != roomToEject.ownerId) {
+    if (user && target && roomToEject && target.id != roomToEject.ownerId) {
       if (roomToEject.admins.find(admin => admin.id == user.id) || user.id == roomToEject.ownerId) {
         await this.chatService.removeUsersFromRoom(roomToEject.id, target.id);
         this.eventEmitter.emit('room.user.kicked', { room: roomToEject, user: target, kicker: user });
@@ -132,7 +132,7 @@ export class WschatService {
     const target = await this.userService.findOne(event.targetId);
     const roomToPromote = await this.chatService.getRoomById(event.roomId);
 
-    if (user) {
+    if (user && target && roomToPromote) {
       if (user.id == roomToPromote.ownerId) {
         this.chatService.addAdminsToRoom(roomToPromote.id, target.id);
         this.updateRoomForUsersInRoom(roomToPromote.id);
@@ -151,7 +151,7 @@ export class WschatService {
     const target = await this.userService.findOne(event.targetId);
     const roomToDemote = await this.chatService.getRoomById(event.roomId);
 
-    if (user) {
+    if (user && target && roomToDemote) {
       if (roomToDemote.ownerId == user.id && roomToDemote.ownerId != target.id) {
         this.chatService.removeAdminsFromRoom(roomToDemote.id, target.id);
         this.updateRoomForUsersInRoom(roomToDemote.id);
@@ -170,7 +170,7 @@ export class WschatService {
     const user = this.onlineUserService.getUser(socketId);
     const room = await this.chatService.getRoomById(message.room.id);
 
-    if (user) {
+    if (user && room) {
       try {
         await this.chatService.newMessage(message);
         this.eventEmitter.emit('room.message.new', { room: room });
@@ -215,15 +215,15 @@ export class WschatService {
     const room = await this.chatService.getRoomById(newRoom.id);
     if (!newRoom.name
         || newRoom.name.length < 3
-        || newRoom.name.length > 20
+        || newRoom.name.length > 30
         || !newRoom.description
         || newRoom.description.length < 3
-        || newRoom.description.length > 100)
+        || newRoom.description.length > 200)
       {
         this.eventEmitter.emit('room.update', {user: user, room: room, success: false, message: 'Les champs ne sont pas valides' });
         return ;
       }
-    if (user) {
+    if (user && room) {
       if (room.ownerId == user.id) {
         await this.chatService.editRoom(newRoom);
         this.eventEmitter.emit('room.update', {user: user, room: newRoom, success: true });
@@ -237,7 +237,6 @@ export class WschatService {
     const user = this.onlineUserService.getUser(socketId);
     const target = await this.userService.findOne(event.userId);
 
-    console.log(event);
     if (user && target && user.id != target.id) {
       if (event.block) {
         await this.userService.blockUser(user.id, target.id);
@@ -253,7 +252,7 @@ export class WschatService {
     let room = await this.chatService.getRoomById(roomId);
 
     for (let user of room.users) {
-      let rooms = await this.chatService.getRoomsFromUser(user.id);
+      const rooms = await this.chatService.getRoomsFromUser(user.id);
       this.sendToUser(user, 'rooms', rooms);
     }
   }
